requirements.txt
discord.py==1.7.3
async-timeout==3.0.1
parlai